32位平台和64位平台对long的输出处理不同。
32位平台long同int相同，但是在64位上，必须加上l。
如果打开-Wall -Werror，会造成在一个平台上编译不通过。

+++ src/libqq/qqmanageinfo.c
@@ -252,7 +252,7 @@ static gint do_get_recent_contact(QQInfo *info, GError **err)
             , info -> vfwebqq -> str);
     gchar *euri = g_uri_escape_string(params, "=", FALSE);
     request_append_msg(req, euri, strlen(euri));
-    g_snprintf(params, 300, "%d", strlen(euri));
+    g_snprintf(params, 300, "%u", (unsigned int)strlen(euri));
     request_add_header(req, "Content-Length", params);
     g_free(euri);
 
@@ -562,7 +562,7 @@ static gint do_get_my_friends(QQInfo *info, GError **err)
             , info -> vfwebqq -> str);
     gchar *euri = g_uri_escape_string(params, "=", FALSE);
     request_append_msg(req, euri, strlen(euri));
-    g_snprintf(params, 300, "%d", strlen(euri));
+    g_snprintf(params, 300, "%u", (unsigned int)strlen(euri));
     request_add_header(req, "Content-Length", params);
     g_free(euri);
 
@@ -793,7 +793,7 @@ static gint do_get_group_name_list_mask(QQInfo *info, GError **err)
             , info -> vfwebqq -> str);
     gchar *euri = g_uri_escape_string(params, "=", FALSE);
     request_append_msg(req, euri, strlen(euri));
-    g_snprintf(params, 300, "%d", strlen(euri));
+    g_snprintf(params, 300, "%u", (unsigned)strlen(euri));
     request_add_header(req, "Content-Length", params);
     g_free(euri);
